requirements.txt
crewai>=0.1.0
pyspark>=3.3.0
python-dotenv>=1.0.1
psycopg2-binary>=2.9.0
PyPDF2>=3.0.0
numpy==1.24.3
pandas==2.0.3
